Thread pool: remove thread ready signal
It seems the workers were sending the "ready" signal before the
parent thread started to wait for it. This lead to timeouts.

The parent thread doesn't really need to wait for the worker
threads to carry on with its own work, so I've decided to remove
the check.

Progress on #49.