RTR Server: poll before writing
Problem:

write() was sometimes failing with EAGAIN when Fort tried to send PDUs
to clients.

Diagnosis:

Obviously, it's because RTR client sockets now ship with O_NONBLOCK
enabled. Fort wants O_NONBLOCK for reading, but not for writing.

This bug was introduced in the previous commit.

Solution:

Make sure the socket is writable (via poll()) before calling write().