Feature test macro review
Stop relying on default features; explicitely define the macros that
enable the symbols we need.

It appears POSIX.1 usage is fairly widespread through the code, and its
minimum required version is 2001.

We also have one GNU extension: getopt_long().

And we also have one GNU extension marked as nonstandard: timegm().
I'm going to try to remove it in a future commit.