Merge the RRDP modules
I don't really have any strong arguments to justify this. Been
progressively convincing myself to do it over time, and at this point I
think it's inevitable.

It's not that much code, most of this stuff doesn't need to be exported,
and the reduced API simplifies the review.

I'm becoming a minimalist, I suppose.