Improve selection of cache RPP fallback
Corner case. Suppose the cache has (for whatever reason) downloaded the
two following URLs separately:

	rsync://a.b.c/d
	rsync://a.b.c/d/e/f

(This might happen if the latter is downloaded in one iteration, then
the former is downloaded the next, and the cleanup timer hasn't kicked
in yet.)

This commit extends the existing priority selection algorithm to this
situation.

(The old way was to choose the most specific URL, which would go on to
lose to a different URL which might have lost to the less specific
version.)

Honestly, this is a micro-correction. It hopefully slightly increases
the chances of the cache fallback being useful in very specific unusual
situations, rather than guarantee it. But it's more consistent, future-
proof and looks more sensible in the tests.