Improve HTTP 204 error message
Looks like Afrinic disabled RRDP, and they're returning 204 instead of
4xx or 5xx.

This was fooling Fort into thinking the file downloaded successfully,
which induced a crappier error message.

But regardless of Afrinic, an empty file is not a valid RPKI file,
so 204 (No Content) should probably be handled as an error anyway.