Synchronize rsync modules instead of RPPs again
rsync-ing every RPP separately is prohibitely expensive when RRDP is
unavailable.

As it turns out, rsync-ing the repository root happens to be an
unwritten standard. Fort was far from the only one doing it, and people
expect it.

This means I will eventually have to come up with a different way to box
RRDP RPPs, as the current implementation induces too many redownloads
when rsync needs to be fall-backed into.

I will have to leave that outside of 1.6.0 however, as I've fixed too
much stuff already, and I need a new release urgently.

Sort of reverts #80, though the flag will remain deleted. I don't think
there's a point in offloading this decision to the user.