Warn if server.address has more than one wildcard
Consider the following three definitions of server.address:

	[a] "address": [ "0.0.0.0" ]
	[b] "address": [ "::" ]
	[c] "address": [ "0.0.0.0", "::" ]

We expect [a] to bind the socket to any IPv4 address,
[b] to bind the socket to any IPv6 address,
and [c] to bind the socket to any IPv4 and IPv6 address.

Right? The BSDs work that way.

But Linux doesn't. For Linux,

	[a] binds to any IPv4 address,
	[b] binds to any IPv4 and IPv6 address,
	and [c] is an error.

But I don't want to override the behavior because some admins are
probably used to it:

	linux$ nc -6lknv :: 7890
	Listening on :: 7890
	Connection received on ::1 52814
	Hello from IPv6!
	Connection received on ::ffff:127.0.0.1 55456
	Hello from IPv4!

Instead, let's print a warning.

Thanks to Benjamim Pinheiro for reporting this quirk.