Add several values to uri_type
In the upcoming commits, the cache is going to need to start managing
the notifications differently from the other downloadables, including
snapshots and deltas.

The logic used to differentiate the different objects from each other
was getting convoluted, and had been naturally converging into the URI
object since the 1.6.0 refactors, IIRC.

So stop beating around the bush and commit to this design. Store the
type in the URI object, and add several values to the uri_type enum.