Automatically download if file is an rsync URL
rsync cannot download into standard output... which means rsync'd files
cannot be elegantly piped as standard output to --mode=print. So either
the rsync has to be done manually by the user... or --mode=print has to
do it internally by itself.

And looking at the code that resulted... I now wish I had gone with the
former option. Because of the long overdue cache refactors, the user
needs to include --tal for this rsync to be compatible with the cache.
This sucks.

As a workaround, Fort will rsync into /tmp if --tal and/or --local-cache
aren't supplied:

	$ fort --mode=print \
		--validation-log.enabled \
		--validation-log.level debug \
		rsync://a.b.c/d/CRL.crl
	...
	May 10 13:32:44 DBG [Validation]: Executing rsync:
	May 10 13:32:44 DBG [Validation]:     rsync
	May 10 13:32:44 DBG [Validation]:     ...
	May 10 13:32:44 DBG [Validation]:     rsync://a.b.c/d/CRL.crl
	May 10 13:32:44 DBG [Validation]:     /tmp/fort-Q7tMhz/CRL.crl
	...
	{
	"tbsCertList": {
		"version": 1,
	...