Move the signal code out of the logging module
The sigaction() code was in logging because it was originally conceived
by the SIGSEGV stack trace printing hack. The SIGPIPE ignorer was also
incidentally moved there at some point, but it has never had anything
to do with logging.

And I'm going to catch more signals in the upcoming commits, so this
really needs to be formalized into its own module.