Move the rsync spawner to a separate process
The fork()s (needed to spawn rsyncs) duplicate Fort's process.

Which is messy in a multithreaded program. Quoting the Linux man page:

> * The child process is created with a single thread—the one that
>   called fork(). The entire virtual address space of the parent is
>   replicated in the child, including the states of mutexes, condition
>   variables, and other pthreads objects. (...)
> * After a fork() in a multithreaded program, the child can safely call
>   only async-signal-safe functions (...) until such time as it calls
>   execve(2).

As far as I can tell, since the forked child was, in fact, careful to
only invoke async-signal-safe functions, this wasn't really a bug.
Still, it wasn't quality architecture either.

Moving the rsync spawner to a dedicated subprocess should stop the forks
from threatening to clash with the multithreading completely.

Relies on the new core loop design, so this won't work properly until
that's implemented.