Clear the cache after a crash
There are many ways in which a mismatching cache index can cause erratic
behavior that's hard to detect. Since the index is written at the end of
the validation cycle, crashing at any point between a cache refresh and
the index write results in a misindexed cache.

Deleting the index after loading it seems to be a reliable way to force
Fort to reset the cache after a crash.

Terminating signals are still a problem, though.